docs(materials): 区块列表调整
