+++ angular.json
@@ -16,7 +16,15 @@
             "main": "src/main.ts",
             "tsConfig": "src/tsconfig.json",
             "polyfills": "src/polyfills.ts",
-            "assets": ["src/assets", "src/_redirects"],
+            "assets": [
+              "src/assets",
+              "src/_redirects",
+              {
+                "glob": "**/*",
+                "input": "./node_modules/@ant-design/icons-angular/src/inline-svg/",
+                "output": "/assets/"
+              }
+            ],
             "styles": ["node_modules/ngx-simplemde/src/index.css", "node_modules/aos/dist/aos.css", "src/styles.less"],
             "scripts": [
               "node_modules/aos/dist/aos.js",
