+++ src/app/routes/materials/list/list.component.html
@@ -1,9 +1,9 @@
-<page-header [title]="'搜索列表'" [tab]="tab">
+<page-header [title]="'物料资源'" [tab]="tab">
   <div class="text-center">
     <nz-input-group nzSearch nzSize="large" [nzAddOnAfter]="addOnAfterBtn" style="width: 520px;">
       <input type="text" nz-input placeholder="请输入" />
       <ng-template #addOnAfterBtn>
-        <button nz-button nzType="primary" nzSize="large" nzSearch>Search</button>
+        <button nz-button nzType="primary" nzSize="large" nzSearch>搜索</button>
       </ng-template>
     </nz-input-group>
   </div>
