+++ src/app/routes/materials/materials.module.ts
@@ -2,9 +2,9 @@ import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { SharedModule } from '@shared/shared.module';
-import { ProListArticlesComponent } from './articles/articles.component';
+import { ProListArticlesComponent } from './blocks/blocks.component';
 import { ProListLayoutComponent } from './list/list.component';
-import { ProListProjectsComponent } from './projects/projects.component';
+import { ProListProjectsComponent } from './scaffolds/scaffolds.component';
 
 const routes: Routes = [
   { path: '', redirectTo: 'list', pathMatch: 'full' },
@@ -12,9 +12,9 @@ const routes: Routes = [
     path: 'list',
     component: ProListLayoutComponent,
     children: [
-      { path: '', redirectTo: 'articles', pathMatch: 'full' },
-      { path: 'articles', component: ProListArticlesComponent, pathMatch: 'full' },
-      { path: 'projects', component: ProListProjectsComponent, pathMatch: 'full' },
+      { path: '', redirectTo: 'blocks', pathMatch: 'full' },
+      { path: 'blocks', component: ProListArticlesComponent, pathMatch: 'full' },
+      { path: 'scaffolds', component: ProListProjectsComponent, pathMatch: 'full' },
     ],
   },
 ];
