+++ src/app/routes/materials/scaffolds/scaffolds.component.html
@@ -51,7 +51,9 @@
           <!-- <avatar-list size="mini">
             <avatar-list-item *ngFor="let m of item.members" [src]="m.avatar" [tips]="m.name"></avatar-list-item>
           </avatar-list> -->
-          <img *ngFor="let m of item.members" src="{{ m.avatar }}" alt="{{ m.name }}" />
+          <div style="display:inline-block; float:right;">
+            <img *ngFor="let m of item.members" src="{{ m.avatar }}" alt="{{ m.name }}" />
+          </div>
         </div>
       </nz-card>
     </nz-list-item>
